Create words in .docx.py
Here is a very niche but useful program for working with .docx documents. I faced with the following task.
Input: Company has package of documentation to be submitted to customer. There are thousands of documents in .docx format with certain  format and style, with drawings in it as well and numerous sections
Task: before submission of these documents some of the information changed and needed to be replace in new version. Namely information in the header of the documents and year of the document on the first page.

So if you need to change text in headers, you are welcome. Program is capable to process number of documents at once and save it in another folder with new name